==> 1-VOI_OriginalandPreQueryTransformation.ipynb <==
# Voi original transformation 

The scope of this script is to assign a trip ID to VOI's original data file and ensure it matches with the pre-query transformation file

import pandas as pd
import requests
import numpy as np
from datetime import datetime, timedelta
import haversine as hs
from haversine import Unit

df_original = pd.read_csv('folder/file.csv')

df = df_original

## 1. Global cleaning 

#1 dropping End-Time & Start Date
df=df.drop(['START_DATE'], axis=1)

#2 renaming columns to match structure of requests.csv
df=df.rename(columns={"START_TIME":"treq","START_LONGITUDE":"origin_x","START_LATITUDE":"origin_y","END_LONGITUDE":"destination_x","END_LATITUDE":"destination_y",})

#3 drop out of bound trips
minLong = 11.8735
minLat = 57.6265
maxLong = 12.0427
maxLat = 57.7712

len1=len(df)
len1
df.drop(df.loc[df['origin_x']<minLong].index, inplace=True)
df.drop(df.loc[df['origin_y']<minLat].index, inplace=True)
df.drop(df.loc[df['destination_x']<minLong].index, inplace=True)
df.drop(df.loc[df['destination_y']<minLat].index, inplace=True)
df.drop(df.loc[df['origin_x']>maxLong].index, inplace=True)
df.drop(df.loc[df['origin_y']>maxLat].index, inplace=True)
df.drop(df.loc[df['destination_x']>maxLong].index, inplace=True)
df.drop(df.loc[df['destination_y']>maxLat].index, inplace=True)
len2=len(df)

print("Rows removed:", len1-len2)


#5. Time and duration: formatting start and end to datetime, incrementing to UTC+1 since we assume that GTFS data is saved in Gothenburg's local time (check with Tobias), AND calculating duration
df['treq'] = pd.to_datetime(df['treq'], format='%Y-%m-%d %H:%M:%S')
df['treq'] = pd.to_datetime(df['treq'].dt.strftime('%Y-%m-%d %H:%M:%S'))

df['END_TIME'] = pd.to_datetime(df['END_TIME'], format='%Y-%m-%d %H:%M:%S')
df['END_TIME'] = pd.to_datetime(df['END_TIME'].dt.strftime('%Y-%m-%d %H:%M:%S'))


df['duration_in_m']=df['END_TIME']-df['treq']
df['duration_in_m']=(df.duration_in_m.dt.total_seconds() / 60)
df

## code to remove rows that have distance and duration under a certain value?


##=============================ASSIGNING ID AFTER ALL CLEANING IS MADE
## Sort oldest trip first, and assign id. ID is being assigned BEFORE the dataset is split by schedule to ensure cardinality if the datasets need to be concatenated 
df.sort_values(by='treq', ascending=True, inplace=True)
df['TRIP_ID']=np.arange(len(df))

#6 split summer and winter trips. IMPORTANT. VOI's ORIGINAL DATA MUST ALSO BE SPLIT TO RETAIN ROW-WISE INDEX IDENTIFIERS
start_date = '2022-06-19' #day summer schedule starts
end_date   = '2022-08-20' #day summer schedule ends

df_summer = df.query('treq > @start_date and treq < @end_date')
df_summer.to_csv('VoiData_CleanedIndexed_Summer.csv', index=False)

#6 dataframe for winter trips
start_date = '2022-06-19' #day summer schedule starts
end_date   = '2022-08-20' #day summer schedule ends

df_winter1 = df.query('treq < @start_date') #all dates before summer
df_winter2 = df.query('treq > @end_date') #all dates after summer
df_winter = pd.concat([df_winter1,df_winter2])
df_winter.to_csv('VoiData_CleanedIndexed_Winter.csv', index=False)

df_winter

## 2. Transformation for Querying

#7 assigning appropriate day of the week 

#probably a runner-up as the world's most inneficient piece of code 

monday = '2022-01-03 '
tuesday = '2022-01-04 '
wednesday = '2022-01-05 '
thursday = '2022-01-06 '
friday = '2022-01-07 '
saturday = '2022-01-08 '
sunday = '2022-01-09 '


df_winter['treq'] = pd.to_datetime(df_winter['treq'], format='%Y-%m-%d %H:%M:%S')
df_winter['treq'] = pd.to_datetime(df_winter['treq'].dt.strftime('%Y-%m-%d %H:%M:%S'))

df_winter['treqDayName']=df_winter['treq'].dt.day_name()
df_winter['time'] = [d.time() for d in df_winter['treq']]

df_winterMon = df_winter.loc[df_winter['treqDayName'] == 'Monday']
df_winterMon['treq'] = pd.to_datetime(df_winterMon['treq'])
df_winterMon['mondayDate'] = monday
df_winterMon['treq'] = df_winterMon['mondayDate'].astype(str) + df_winterMon['time'].astype(str)
df_winterMon=df_winterMon.drop(['treqDayName','time','mondayDate'], axis=1)

df_winterTue = df_winter.loc[df_winter['treqDayName'] == 'Tuesday']
df_winterTue['treq'] = pd.to_datetime(df_winterTue['treq'])
df_winterTue['tuesdayDate'] = tuesday
df_winterTue['treq'] = df_winterTue['tuesdayDate'].astype(str) + df_winterTue['time'].astype(str)
df_winterTue=df_winterTue.drop(['treqDayName','time','tuesdayDate'], axis=1)

df_winterWed= df_winter.loc[df_winter['treqDayName'] == 'Wednesday']
df_winterWed['treq'] = pd.to_datetime(df_winterWed['treq'])
df_winterWed['wednesdayDate'] = wednesday
df_winterWed['treq'] = df_winterWed['wednesdayDate'].astype(str) + df_winterWed['time'].astype(str)
df_winterWed=df_winterWed.drop(['treqDayName','time','wednesdayDate'], axis=1)

df_winterThur= df_winter.loc[df_winter['treqDayName'] == 'Thursday']
df_winterThur['treq'] = pd.to_datetime(df_winterThur['treq'])
df_winterThur['thursdayDate'] = thursday
df_winterThur['treq'] = df_winterThur['thursdayDate'].astype(str) + df_winterThur['time'].astype(str)
df_winterThur=df_winterThur.drop(['treqDayName','time','thursdayDate'], axis=1)

df_winterFri= df_winter.loc[df_winter['treqDayName'] == 'Friday']
df_winterFri['treq'] = pd.to_datetime(df_winterFri['treq'])
df_winterFri['fridayDate'] = friday
df_winterFri['treq'] = df_winterFri['fridayDate'].astype(str) + df_winterFri['time'].astype(str)
df_winterFri=df_winterFri.drop(['treqDayName','time','fridayDate'], axis=1)

df_winterSat= df_winter.loc[df_winter['treqDayName'] == 'Saturday']
df_winterSat['treq'] = pd.to_datetime(df_winterSat['treq'])
df_winterSat['saturdayDate'] = saturday
df_winterSat['treq'] = df_winterSat['saturdayDate'].astype(str) + df_winterSat['time'].astype(str)
df_winterSat=df_winterSat.drop(['treqDayName','time','saturdayDate'], axis=1)

df_winterSun= df_winter.loc[df_winter['treqDayName'] == 'Sunday']
df_winterSun['treq'] = pd.to_datetime(df_winterSun['treq'])
df_winterSun['sundayDate'] = sunday
df_winterSun['treq'] = df_winterSun['sundayDate'].astype(str) + df_winterSun['time'].astype(str)
df_winterSun=df_winterSun.drop(['treqDayName','time','sundayDate'], axis=1)


df_winter = pd.concat([df_winterMon,df_winterTue,df_winterWed,df_winterThur,df_winterFri,df_winterSat,df_winterSun])
df_winter.sort_values(by='TRIP_ID', ascending=True, inplace=True)

#7 assigning appropriate day of the week 

#probably a runner-up as the world's most inneficient piece of code 
monday = '2022-06-27 '
tuesday = '2022-06-28 '
wednesday = '2022-06-29 '
thursday = '2022-06-30 '
friday = '2022-07-01 '
saturday = '2022-07-02 '
sunday = '2022-07-03 '


df_summer['treq'] = pd.to_datetime(df_summer['treq'], format='%Y-%m-%d %H:%M:%S')
df_summer['treq'] = pd.to_datetime(df_summer['treq'].dt.strftime('%Y-%m-%d %H:%M:%S'))

df_summer['treqDayName']=df_summer['treq'].dt.day_name()
df_summer['time'] = [d.time() for d in df_summer['treq']]

df_summerMon = df_summer.loc[df_summer['treqDayName'] == 'Monday']
df_summerMon['treq'] = pd.to_datetime(df_summerMon['treq'])
df_summerMon['mondayDate'] = monday
df_summerMon['treq'] = df_summerMon['mondayDate'].astype(str) + df_summerMon['time'].astype(str)
df_summerMon=df_summerMon.drop(['treqDayName','time','mondayDate'], axis=1)

df_summerTue = df_summer.loc[df_summer['treqDayName'] == 'Tuesday']
df_summerTue['treq'] = pd.to_datetime(df_summerTue['treq'])
df_summerTue['tuesdayDate'] = tuesday
df_summerTue['treq'] = df_summerTue['tuesdayDate'].astype(str) + df_summerTue['time'].astype(str)
df_summerTue=df_summerTue.drop(['treqDayName','time','tuesdayDate'], axis=1)

df_summerWed= df_summer.loc[df_summer['treqDayName'] == 'Wednesday']
df_summerWed['treq'] = pd.to_datetime(df_summerWed['treq'])
df_summerWed['wednesdayDate'] = wednesday
df_summerWed['treq'] = df_summerWed['wednesdayDate'].astype(str) + df_summerWed['time'].astype(str)
df_summerWed=df_summerWed.drop(['treqDayName','time','wednesdayDate'], axis=1)

df_summerThur= df_summer.loc[df_summer['treqDayName'] == 'Thursday']
df_summerThur['treq'] = pd.to_datetime(df_summerThur['treq'])
df_summerThur['thursdayDate'] = thursday
df_summerThur['treq'] = df_summerThur['thursdayDate'].astype(str) + df_summerThur['time'].astype(str)
df_summerThur=df_summerThur.drop(['treqDayName','time','thursdayDate'], axis=1)

df_summerFri= df_summer.loc[df_summer['treqDayName'] == 'Friday']
df_summerFri['treq'] = pd.to_datetime(df_summerFri['treq'])
df_summerFri['fridayDate'] = friday
df_summerFri['treq'] = df_summerFri['fridayDate'].astype(str) + df_summerFri['time'].astype(str)
df_summerFri=df_summerFri.drop(['treqDayName','time','fridayDate'], axis=1)

df_summerSat= df_summer.loc[df_summer['treqDayName'] == 'Saturday']
df_summerSat['treq'] = pd.to_datetime(df_summerSat['treq'])
df_summerSat['saturdayDate'] = saturday
df_summerSat['treq'] = df_summerSat['saturdayDate'].astype(str) + df_summerSat['time'].astype(str)
df_summerSat=df_summerSat.drop(['treqDayName','time','saturdayDate'], axis=1)

df_summerSun= df_summer.loc[df_summer['treqDayName'] == 'Sunday']
df_summerSun['treq'] = pd.to_datetime(df_summerSun['treq'])
df_summerSun['sundayDate'] = sunday
df_summerSun['treq'] = df_summerSun['sundayDate'].astype(str) + df_summerSun['time'].astype(str)
df_summerSun=df_summerSun.drop(['treqDayName','time','sundayDate'], axis=1)


df_summer = pd.concat([df_summerMon,df_summerTue,df_summerWed,df_summerThur,df_summerFri,df_summerSat,df_summerSun])
df_summer.sort_values(by='TRIP_ID', ascending=True, inplace=True)
df_summer

#8 export to CSV 
df_winter.to_csv('Q/georequestsWinter.csv', index=False)
df_summer.to_csv('Q/georequestsSummer.csv', index=False)
