==> 2-OTPQuerying.ipynb <==
#### Sample API query
http://localhost:8080/otp/routers/default/plan?fromPlace=57.68845%2C11.93338&toPlace=57.67121%2C11.93156&time=1%3A02pm&date=11-10-2022&mode=TRANSIT%2CWALK&maxWalkDistance=5000&arriveBy=false


import pandas as pd
import requests 
from mainCopy import make_query, parse_OTP_response
from utils import plot
import numpy as np

## 1. Querying

PATH = 'Q/georequestsSummer.csv'
OTP_API = "http://localhost:8080/otp/routers/default/plan"

df = pd.read_csv(PATH, index_col=[0], skiprows=[i for i in range(1,70001)])# skipping first n+1 (since n does not include the given row) rows since they were successful in the previous run
#df = pd.read_csv(PATH)  # load the csv #index_col=[0]
df.treq=pd.to_datetime(df.treq)

print("Total trips in dataframe:", len(df))
df

### Querying API
*Ensure Grizzly server is running*

java -Xmx4G -jar otp.jar --load inputFiles

row = df.head(1).squeeze()
query = make_query(row)
r = requests.get(OTP_API, params=query) #route r as response from API
route = parse_OTP_response(r.json()) #route r parsed into JSON
route

routeList = []

for index, row in df.iterrows():
    query = make_query(row)
    r = requests.get(OTP_API, params=query) #route r as response from API
    route = parse_OTP_response(r.json()) #route r parsed into JSON
    routList=routeList.append(route)
    
dfResp = pd.DataFrame(routeList)
dfResp.to_csv('R/responses.csv')


dfResp['success'].value_counts()[False]
#dfResp
