==> 3-PostQueryReformatting.ipynb <==
import pandas as pd
from utils import plot
import numpy as np
import datetime

## 1. Merge successful files & Convert epoch to datetime

df = pd.read_csv('R/responses.csv')
df_o = pd.read_csv ('VoiData_CleanedIndexed_Summer.csv')


# REMEMBER! treq_r is not the actual date of a trip. For real date use start_time

## 1.  Fill nan values and convert all exponential 'times' back to int

df = df.fillna(0) #fill all NaN values

def convert(x):
    try:
        return int(float(x))
    except Exception:
        return x

## CONVERT ALL TIME FIELDS TO INT
df['treq_r']=df['treq_r'].apply(convert) 
df['startTime']=df['startTime'].apply(convert) 
df['endTime']=df['endTime'].apply(convert) 

## Convert epoch time to normal date time 

df['treq_r']=df['treq_r'].astype('int64') # using int64 because of integer size
df['startTime']=df['startTime'].astype('int64') # using int64 because of integer size
df['endTime']=df['endTime'].astype('int64') # using int64 because of integer size

df['treq_r'] = df['treq_r']/1000.0 # division by 1000. Not sure why yet, but it's what worked. Probably because milliseconds are originally included and therefore extra 0s
df['startTime']=df['startTime']/1000.0 # division by 1000. Not sure why yet, but it's what worked. Probably because milliseconds are originally included and therefore extra 0s
df['endTime']=df['endTime']/1000.0 # division by 1000. Not sure why yet, but it's what worked. Probably because milliseconds are originally included and therefore extra 0s


f = lambda time : datetime.datetime.fromtimestamp(time).strftime('%Y-%m-%d %H:%M:%S.%f') #function to convert epoch to datetime 
df['treq_r']= df.apply(lambda x: f(x.treq_r), axis=1) #calling function on a specific column
df['startTime']= df.apply(lambda x: f(x.startTime), axis=1) #calling function on a specific column
df['endTime']= df.apply(lambda x: f(x.endTime), axis=1) #calling function on a specific column 

df.head(1) #This is from the response dataset. Check with df.head() below to ensure that timings tally 

## 2. Assign IDs

# Summer IDs start  till 772142. 
df['TRIP_ID'] = df.index+529866

df.to_csv('R/conc/otpSummerAll_Trans.csv', index=False)
df
